pyproject.toml
[project]
name = "touying"
version = "0.11.1"
description = "Export presentation slides in various formats for Touying"
readme = "README.md"
keywords = ["presentation", "slides", "export"]
license = {text = "MIT"}
authors = [
    {name = "OrangeX4", email = "318483724@qq.com"}
]
dependencies = [
    "typst>=0.11.1",
    "python-pptx>=0.6",
    "pillow",
    "jinja2",
    "scour",
    "argparse",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
]

[project.urls]
Home = "https://github.com/touying-typ/touying-exporter"

[project.scripts]
touying = "touying.cli:main"

[tool.setuptools]
packages = {find = {exclude = ["tests"]}}
py-modules = ["touying"]

[tool.setuptools.package-data]
touying = ["*.j2"]